refactor(Name): 패키지 이동 (domain -> domain.player)
